backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0

# LLM
google-generativeai==0.8.0

# Embeddings & Vector Store
sentence-transformers==3.0.0
faiss-cpu==1.8.0
numpy==1.26.0

# Document Processing
PyMuPDF==1.24.0
pytesseract==0.3.10
Pillow==10.4.0

# PDF Creation (for sample docs)
reportlab==4.2.0

# Utilities
python-dotenv==1.0.0
